+++ TowersOfHanoi/TowersOfHanoi.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2010
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TowersOfHanoi", "TowersOfHanoi\TowersOfHanoi.csproj", "{AD83AEF0-D90D-449C-81C5-8F6A7A78DF39}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{AD83AEF0-D90D-449C-81C5-8F6A7A78DF39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD83AEF0-D90D-449C-81C5-8F6A7A78DF39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD83AEF0-D90D-449C-81C5-8F6A7A78DF39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD83AEF0-D90D-449C-81C5-8F6A7A78DF39}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BCC21E61-5903-4FA3-BC81-49824DE59E86}
+	EndGlobalSection
+EndGlobal
