+++ TowersOfHanoi/TowersOfHanoi/bin/Debug/netcoreapp2.0/TowersOfHanoi.runtimeconfig.json
@@ -1,0 +1,9 @@
+{
+  "runtimeOptions": {
+    "tfm": "netcoreapp2.0",
+    "framework": {
+      "name": "Microsoft.NETCore.App",
+      "version": "2.0.0"
+    }
+  }
+}
